requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
langdetect
openai

